fixing the array type in python

+++ tests/test_populate_galaxies_wrapper.py
@@ -19,20 +19,22 @@ with patch("populate_galaxies_wrapper.populate_galaxies") as mock_populate:
 
 def test_populate_galaxies():
     # Create mock halo and subsample dictionaries with test data
-    h_mass = np.array([1e13, 2e13, 3e13]),  # Halo masses
-    h_x = np.array([0, 10, 20]),  # x coordinates
-    h_y = np.array([0, 10, 20]),  # y coordinates
-    h_z = np.array([0, 10, 20]),  # z coordinates
-    h_sigma = np.array([0.2, 0.3, 0.4]),  # sigma values
-    h_velocity = np.array([10, 20, 30])  # Velocities
+    h_mass = np.array([1e13, 2e13, 3e13], dtype=np.float32),  # Halo masses
+    h_x = np.array([0, 10, 20], dtype=np.float32),  # x coordinates
+    h_y = np.array([0, 10, 20], dtype=np.float32),  # y coordinates
+    h_z = np.array([0, 10, 20], dtype=np.float32),  # z coordinates
+    h_sigma = np.array([0.2, 0.3, 0.4], dtype=np.float32),  # sigma values
+    h_velocity = np.array([10, 20, 30], dtype=np.float32)  # Velocities
 
-    s_mass = np.array([1e14, 2e14, 3e14]),  # Subsample masses
-    s_host_velocity = np.array([1, 2, 3]),  # Host velocities
-    s_n_particles = np.array([100, 200, 300]),  # Number of particles
-    s_x = np.array([100, 110, 120]),  # x coordinates
-    s_y = np.array([100, 110, 120]),  # y coordinates
-    s_z = np.array([100, 110, 120]),  # z coordinates
-    s_velocity = np.array([10, 20, 30])  # Velocities
+    # Subsample masses
+    s_mass = np.array([1e14, 2e14, 3e14], dtype=np.float32),
+    s_host_velocity = np.array([1, 2, 3], dtype=np.float32),  # Host velocities
+    # Number of particles
+    s_n_particles = np.array([100, 200, 300], dtype=np.int32),
+    s_x = np.array([100, 110, 120], dtype=np.float32),  # x coordinates
+    s_y = np.array([100, 110, 120], dtype=np.float32),  # y coordinates
+    s_z = np.array([100, 110, 120], dtype=np.float32),  # z coordinates
+    s_velocity = np.array([10, 20, 30], dtype=np.float32)  # Velocities
 
     # Mock HOD parameters
     # [lnMcut, sigma, lnM1, kappa, alpha, alpha_c, alpha_s]
